Data Project Management/Data Project 1 - Meta/G4_DP1-main/requirements.txt
keyboard~=0.13.5
Faker~=11.3.0
kafka-python~=2.0.2
geopy~=2.2.0
pandas~=1.4.0
folium~=0.12.1.post1
numpy~=1.22.1
altair~=4.2.0
SQLAlchemy~=1.4.31
psycopg2~=2.9.3
Flask~=2.0.2
pykafka~=2.8.0